refactor(controller) : 패키지 이동 및, toEntity 사용
